#154: Fix issue with the API when served via WSGI
