+++ test/icat/conftest.py
@@ -8,7 +8,10 @@ from icat.query import Query
 import pytest
 
 from datagateway_api.common.config import config
-from datagateway_api.src.main import create_api_endpoints, create_app_infrastructure
+from datagateway_api.src.api_start_utils import (
+    create_api_endpoints,
+    create_app_infrastructure,
+)
 from test.icat.test_query import prepare_icat_data_for_assertion
